services/woo/requirements.txt
nameko
nameko_sentry
circuitbreaker
woocommerce
nameko_sqlalchemy
alembic==1.0.10